Added webview/browser feature
added the webview feature that uses the link textbox to open it in a
webview or browser if it deems it necessary (used/then edited the code
that our teacher gave out in the webview tutorial).  Just note, that as
it is right now, the user must type in the http:// part (else it doesn't
seem to work).  If this is a big problem let me know, but I don't think
it is.