Added a checknetwork class....
Added a CheckNetwork class, which does as it sounds (it checks if
internet is available).  To test this I just added a call on the
CreateGroupActivity, which will output if network is connected (using
Toast).  To check if it isn't using an emulator, turn Airplane mode on
and then create a group (it will display Network False).  This check
happens when CreateGroupActivity first runs, which obviously isn't what
we are going to do, for now this just shows that it works, and then
tomorrow (6/6/16) we will work out where all the places the check should
be placed.